Upstream LiveResolverStore
