+++ packages/react-relay/relay-hooks/useFragmentInternal_EXPERIMENTAL.js
@@ -537,7 +537,7 @@ hook useFragmentInternal_EXPERIMENTAL(
     if (suspendingLiveResolvers != null && suspendingLiveResolvers.length > 0) {
       throw Promise.all(
         suspendingLiveResolvers.map(({liveStateID}) => {
-          // $FlowFixMe[prop-missing] This is expected to be a LiveResolverStore
+          // $FlowFixMe[prop-missing] This is expected to be a RelayModernStore
           return environment.getStore().getLiveResolverPromise(liveStateID);
         }),
       );
