+++ packages/relay-runtime/store/__tests__/ClientEdgeToClientObject-test.js
@@ -16,11 +16,11 @@ import type {ClientEdgeToClientObjectTest3Query$data} from './__generated__/Clie
 const {RelayFeatureFlags, commitLocalUpdate} = require('relay-runtime');
 const RelayNetwork = require('relay-runtime/network/RelayNetwork');
 const {graphql} = require('relay-runtime/query/GraphQLTag');
-const LiveResolverStore = require('relay-runtime/store/experimental-live-resolvers/LiveResolverStore.js');
 const RelayModernEnvironment = require('relay-runtime/store/RelayModernEnvironment');
 const {
   createOperationDescriptor,
 } = require('relay-runtime/store/RelayModernOperationDescriptor');
+const RelayModernStore = require('relay-runtime/store/RelayModernStore.js');
 const RelayRecordSource = require('relay-runtime/store/RelayRecordSource');
 const {
   disallowConsoleErrors,
@@ -85,7 +85,7 @@ describe.each([true, false])(
       `;
 
       const operation = createOperationDescriptor(FooQuery, {});
-      const store = new LiveResolverStore(source, {
+      const store = new RelayModernStore(source, {
         gcReleaseBufferSize: 0,
       });
 
@@ -139,7 +139,7 @@ describe.each([true, false])(
       `;
 
       const operation = createOperationDescriptor(FooQuery, {});
-      const store = new LiveResolverStore(source, {
+      const store = new RelayModernStore(source, {
         gcReleaseBufferSize: 0,
       });
 
@@ -229,7 +229,7 @@ describe.each([true, false])(
       `;
 
       const operation = createOperationDescriptor(FooQuery, {});
-      const liveStore = new LiveResolverStore(source, {
+      const liveStore = new RelayModernStore(source, {
         gcReleaseBufferSize: 0,
       });
