+++ packages/relay-runtime/store/__tests__/resolvers/LiveResolvers-test.js
@@ -22,11 +22,11 @@ const {
   GLOBAL_STORE,
   resetStore,
 } = require('relay-runtime/store/__tests__/resolvers/ExampleExternalStateStore');
-const LiveResolverStore = require('relay-runtime/store/experimental-live-resolvers/LiveResolverStore');
 const RelayModernEnvironment = require('relay-runtime/store/RelayModernEnvironment');
 const {
   createOperationDescriptor,
 } = require('relay-runtime/store/RelayModernOperationDescriptor');
+const RelayModernStore = require('relay-runtime/store/RelayModernStore');
 const RelayRecordSource = require('relay-runtime/store/RelayRecordSource');
 const {
   disallowConsoleErrors,
@@ -81,7 +81,7 @@ test('unsubscribe happens when record is updated due to missing data', () => {
     `,
     {},
   );
-  const store = new LiveResolverStore(source, {
+  const store = new RelayModernStore(source, {
     gcReleaseBufferSize: 0,
   });
   const environment = new RelayModernEnvironment({
@@ -146,7 +146,7 @@ test('Updates can be batched', () => {
     {},
   );
   const log = jest.fn();
-  const store = new LiveResolverStore(source, {
+  const store = new RelayModernStore(source, {
     gcReleaseBufferSize: 0,
     log,
   });
@@ -252,7 +252,7 @@ test('Errors thrown during _initial_ read() are caught as resolver errors', () =
     `,
     {},
   );
-  const store = new LiveResolverStore(source, {
+  const store = new RelayModernStore(source, {
     gcReleaseBufferSize: 0,
   });
   const environment = new RelayModernEnvironment({
@@ -288,7 +288,7 @@ test('Errors thrown during read() _after update_ are caught as resolver errors',
     `,
     {},
   );
-  const store = new LiveResolverStore(source, {
+  const store = new RelayModernStore(source, {
     gcReleaseBufferSize: 0,
   });
   const environment = new RelayModernEnvironment({
@@ -335,4 +335,85 @@ test('Errors thrown during read() _after update_ are caught as resolver errors',
   expect(finalSnapshot.relayResolverErrors).toEqual([]);
   const finalData: $FlowExpectedError = finalSnapshot.data;
   expect(finalData.counter_throws_when_odd).toBe(2);
+});
+
+test('Reflects optimistic updates, and reverts of optimistic updates', () => {
+  const source = RelayRecordSource.create({
+    'client:root': {
+      __id: 'client:root',
+      __typename: '__Root',
+      me: {__ref: '4'},
+    },
+    '4': {
+      __id: '4',
+      __typename: 'User',
+      name: 'Alice',
+    },
+  });
+  const operation = createOperationDescriptor(
+    graphql`
+      query LiveResolversTestOptimisticUpdatesQuery {
+        me {
+          greeting
+        }
+      }
+    `,
+    {},
+  );
+
+  const store = new RelayModernStore(source, {
+    gcReleaseBufferSize: 0,
+  });
+  const environment = new RelayModernEnvironment({
+    network: RelayNetwork.create(jest.fn()),
+    store,
+  });
+
+  const snapshot = environment.lookup(operation.fragment);
+
+  // When we start we greet Alice
+  expect(snapshot.data).toEqual({me: {greeting: 'Hello, Alice!'}});
+
+  const handler = jest.fn<[Snapshot], void>();
+  environment.subscribe(snapshot, handler);
+
+  const mutationSub = environment.applyMutation({
+    operation: createOperationDescriptor(
+      graphql`
+        mutation LiveResolversTestOptimisticUpdatesMutation(
+          $input: ActorNameChangeInput!
+        ) {
+          actorNameChange(input: $input) {
+            actor {
+              name
+            }
+          }
+        }
+      `,
+      {input: {newName: 'Bob'}},
+    ),
+    response: {
+      actorNameChange: {
+        actor: {
+          __typename: 'User',
+          id: '4',
+          name: 'Bob',
+        },
+      },
+    },
+    updater: null,
+  });
+
+  // In optimistic state we greet Bob
+  expect(handler.mock.calls.length).toBe(1);
+  expect(handler.mock.calls[0][0].data).toEqual({
+    me: {greeting: 'Hello, Bob!'},
+  });
+
+  mutationSub.dispose();
+  // When the optimistic update is disposed we greet Alice again
+  expect(handler.mock.calls.length).toBe(2);
+  expect(handler.mock.calls[1][0].data).toEqual({
+    me: {greeting: 'Hello, Alice!'},
+  });
 });
