--- packages/relay-runtime/store/experimental-live-resolvers/README.md
@@ -1,5 +1,0 @@
-# [EXPERIMENTAL] Live Resolvers
-
-This directory contains an experimental implementation of Store to explore a potential implementation of Live Resolvers, a variant of [Relay Resolvers](https://relay.dev/docs/next/guides/relay-resolvers/) that supports resolvers which expose data from another data source, such as a legacy state management library.
-
-This code is not production ready and is being written purely as an exploration.
