+++ packages/relay-runtime/store/live-resolvers/LiveResolverCache.js
@@ -30,7 +30,7 @@ import type {
   GetDataForResolverFragmentFn,
   ResolverCache,
 } from '../ResolverCache';
-import type LiveResolverStore from './LiveResolverStore';
+import type RelayModernStore from './RelayModernStore';
 import type {LiveState} from 'relay-runtime';
 
 const recycleNodesInto = require('../../util/recycleNodesInto');
@@ -91,13 +91,13 @@ class LiveResolverCache implements ResolverCache {
   _resolverIDToRecordIDs: Map<ResolverID, Set<DataID>>;
   _recordIDToResolverIDs: Map<DataID, Set<ResolverID>>;
   _getRecordSource: () => MutableRecordSource;
-  _store: LiveResolverStore;
+  _store: RelayModernStore;
   _handlingBatch: boolean; // Flag indicating that Live Resolver updates are being batched.
   _liveResolverBatchRecordSource: ?MutableRecordSource; // Lazily created record source for batched Live Resolver updates.
 
   constructor(
     getRecordSource: () => MutableRecordSource,
-    store: LiveResolverStore,
+    store: RelayModernStore,
   ) {
     this._resolverIDToRecordIDs = new Map();
     this._recordIDToResolverIDs = new Map();
